refactor: consolidate imports in useComparisonData and useEditPage hooks for improved readability
