+++ src/features/comparison/hooks/useComparisonData.ts
@@ -4,8 +4,7 @@ import { useQueries } from '@tanstack/react-query';
 
 import { SharedInfraContext } from '@/contexts';
 
-import { useLoadProfile } from '@/features/profiles/hooks/useLoadProfile';
-import { useLoadProfileSettings } from '@/features/profiles/hooks/useLoadProfileSettings';
+import { useLoadProfile, useLoadProfileSettings } from '@/features/profiles';
 import { fetchAndBuildPreview } from '@/features/resources';
 
 export const useComparisonData = (resourceIds: string[]) => {
